Crop-head-counter/requirements.txt
Pillow==10.4.0
torch==2.4.1
torchvision==0.19.1
